[Register] Moved save of 'info_header_draw' in NIMPHS_Scene
